Possible solution for #35
As mentioned in the comments, github API allows to use a `since:date` key in searches that shows issues updated since that date, but not the other way around. In the solution above I assume that among the results there is one that *has not* been updated since 14 days.

I made few tests and it seems to do the trick.